+++ api/core/tools/provider/builtin/chart/tools/bar.py
@@ -4,7 +4,6 @@ from typing import Any, Union
 import matplotlib.pyplot as plt
 
 from core.tools.entities.tool_entities import ToolInvokeMessage
-from core.tools.provider.builtin.chart.mpl_setting import font_properties
 from core.tools.tool.builtin_tool import BuiltinTool
 
 
@@ -29,13 +28,11 @@ class BarChartTool(BuiltinTool):
             if len(axis) != len(data):
                 axis = None
 
-        plt.rcParams["font.family"] = font_properties.get_name()
-
         flg, ax = plt.subplots(figsize=(10, 8))
 
         if axis:
             axis = [label[:10] + "..." if len(label) > 10 else label for label in axis]
-            ax.set_xticklabels(axis, rotation=45, ha="right", fontproperties=font_properties)
+            ax.set_xticklabels(axis, rotation=45, ha="right")
             ax.bar(axis, data)
         else:
             ax.bar(range(len(data)), data)
