+++ api/core/tools/provider/builtin/chart/tools/pie.py
@@ -4,7 +4,6 @@ from typing import Any, Union
 import matplotlib.pyplot as plt
 
 from core.tools.entities.tool_entities import ToolInvokeMessage
-from core.tools.provider.builtin.chart.mpl_setting import font_properties
 from core.tools.tool.builtin_tool import BuiltinTool
 
 
@@ -26,8 +25,6 @@ class PieChartTool(BuiltinTool):
         else:
             data = [float(i) for i in data]
 
-        plt.rcParams["font.family"] = font_properties.get_name()
-
         flg, ax = plt.subplots()
 
         if categories:
@@ -36,9 +33,9 @@ class PieChartTool(BuiltinTool):
                 categories = None
 
         if categories:
-            ax.pie(data, labels=categories, textprops={"fontproperties": font_properties})
+            ax.pie(data, labels=categories)
         else:
-            ax.pie(data, textprops={"fontproperties": font_properties})
+            ax.pie(data)
 
         buf = io.BytesIO()
         flg.savefig(buf, format="png")
